python-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.50.0
supabase==2.7.0
pandas==2.2.2
numpy>=1.26.0
openpyxl==3.1.5
python-multipart==0.0.9
pydantic==2.9.0
httpx==0.27.2
